scene setup in each photon scene
playground setup in photon scenes based on the number of players the playground is scaled for each. also tried to work on source control with git to hopefully have less commits that fully break the game.  also why i made my own branch so if it breaks it doesn't break main